Fix MapUtils.getIntValue(Map, K, Function) returning a byte (#704)
* fix MapUtils.getIntValue(Map, K, Function) returning a byte

* drive getShortValue function overload past byte range

+++ src/main/java/org/apache/commons/collections4/MapUtils.java
@@ -738,7 +738,7 @@ public class MapUtils {
      */
     public static <K> int getIntValue(final Map<? super K, ?> map, final K key,
             final Function<K, Integer> defaultFunction) {
-        return applyDefaultFunction(map, key, MapUtils::getInteger, defaultFunction, 0).byteValue();
+        return applyDefaultFunction(map, key, MapUtils::getInteger, defaultFunction, 0).intValue();
     }
 
     /**
